autosave photos in profile edit
The expected behavior:
When you upload a photo, it saves the data from the whole page but does not take you to the profile page (which is what happens when you click save). If you navigate directly to your profile without clicking save, you should still see the photo you uploaded on your profile.